+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/data/LongGaugeData.java
@@ -9,9 +9,10 @@ import com.google.auto.value.AutoValue;
 import java.util.Collection;
 import javax.annotation.concurrent.Immutable;
 
+/** {@link GaugeData} recorded uses {@code long}s. */
 @Immutable
 @AutoValue
-public abstract class LongGaugeData implements Data<LongPointData> {
+public abstract class LongGaugeData implements GaugeData<LongPointData> {
   public static LongGaugeData create(Collection<LongPointData> points) {
     return new AutoValue_LongGaugeData(points);
   }