+++ sdk/testing/src/main/java/io/opentelemetry/sdk/testing/time/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package io.opentelemetry.sdk.testing.time;
+
+import javax.annotation.ParametersAreNonnullByDefault;